requirements.txt
whoop
python-dotenv
pandas